- 시리즈 상세 정보 반환 - 시리즈 보유중인 가게 정보 반환

+++ src/main/java/com/example/toonieproject/dto/Book/AddBookRequest.java
@@ -14,7 +14,7 @@ public class AddBookRequest {
 
     private Long storeId;
 
-    private Double rentalPrice;
+    private Integer rentalPrice;
 
     private Integer seriesNum;
